backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
supabase==2.3.4
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Broker integration
alpaca-trade-api==3.0.2

# Market data
yfinance==0.2.36
pandas==2.2.0
numpy==1.26.4

# Sentiment analysis
transformers==4.37.2
torch==2.2.0
feedparser==6.0.11
praw==7.7.1

# LLM
anthropic==0.18.1

# Email
resend==0.8.0

# Utilities
python-dotenv==1.0.1
httpx==0.26.0
pydantic==2.6.1
pydantic-settings==2.1.0

# Task scheduling
apscheduler==3.10.4

# Security
cryptography==42.0.2
